requirements.txt
python-dotenv
google-cloud-aiplatfor
pyyaml